All notification backend completed.

+++ app/Http/Controllers/Admin/PostController.php
@@ -12,6 +12,10 @@ use Image;
 use Validator;
 use Session;
 use Storage;
+use App\Faculty;
+use App\User;
+use Notification;
+use App\Notifications\PostNotification;
 
 class PostController extends Controller
 {   
@@ -59,7 +63,18 @@ class PostController extends Controller
             'content' => 'required|max:10000',
             'tags' => 'required|array|max:20',
             'tags.*' => 'required|max:255|string',
-            'image' => 'nullable|file|image|max:5120'
+            'image' => 'nullable|file|image|max:5120',
+
+            'facultyn' => 'sometimes|required|array|max:12',
+            'facultyn.*' => 'sometimes|required|string',
+            'start_rollno' => 'sometimes|array|max:12',
+            'start_rollno.*' =>'sometimes|required_with:facultyn.*|integer|min:0|max:400',
+            'end_rollno' => 'sometimes|required_with:start_rollno|array|max:12',
+            'end_rollno.*' => 'sometimes|required_with:facultyn.*|integer|min:0|max:400',
+            'year' => 'sometimes|required|array|max:12',
+            'year.*' => 'sometimes|required_with:facultyn.*|integer',
+            'ind_rollno' => 'sometimes|nullable|array|max:40',
+            'ind_rollno.*' => 'sometimes|nullable|alpha_num|size:10'
         ])->validate();
 
         $post= new Post;
@@ -108,8 +123,9 @@ class PostController extends Controller
                     Storage::makeDirectory($dir, 0775, true);
                 }
 
-               Image::make($image)->resize(600, 350,
+               Image::make($image)->resize(720, 720,
                                         function ($constraint) {
+                                             $constraint->aspectRatio();
                                             $constraint->upsize();
                                         })->save(public_path($path));
 
@@ -128,8 +144,49 @@ class PostController extends Controller
 
 
             if($post->tags()->sync($tag_ids, false))
-            {
+            {   
                 Session::flash('success', 'successfully added new post to discussions and forums'.$msg);
+                //Notification Part --start
+                if($request->submit == 'Yes')
+                {
+                    $roll_no=[];
+                    $college = strtoupper(session('tenant'));
+                    for($i=0; $i<count($request->facultyn); $i++)
+                    {  
+                        if( $request->end_rollno[$i] < $request->start_rollno[$i] )
+                               $request->end_rollno[$i]= $request->start_rollno[$i];
+                           
+                        $faculties=[]; 
+                        if($request->facultyn[$i] == 'All')
+                            $faculties = Faculty::pluck('name');
+                         else
+                         $faculties[] =  Faculty::findOrFail(2)->value('name');
+                       
+                        for($k=0; $k<count($faculties); $k++)
+                        {
+                            for($j=$request->start_rollno[$i]; $j<=$request->end_rollno[$i]; $j++ )
+                            {   
+                                $no = str_pad($j, 3, '0', STR_PAD_LEFT);
+                                $roll_no[]=$college.$no.$faculties[$k].$request->year[$i];
+                            }
+                        }
+                    }      
+
+                    $users = User::whereIn('roll_no', $roll_no)->get();
+                     
+                     if($request->ind_rollno) 
+                     {  
+                        foreach($request->ind_rollno as $roll) 
+                            $i_rollno[] = $college.$roll;
+                        
+                        if($i_users = User::whereIn('roll_no', $i_rollno)->get())
+                            (count($users)) ? $users->merge($i_users) : $users = $i_users;
+                                           
+                     }
+                    
+                    Notification::send($users->unique(), new PostNotification($post));              
+                }
+                //Notification Part --end
                 return redirect()->route('posts.show', $post->slug);
             }
             else
@@ -241,8 +298,9 @@ class PostController extends Controller
                         Storage::makeDirectory($dir, 0775, true);
                     }
 
-                   Image::make($image)->resize(600, 350,
+                   Image::make($image)->resize(720, 720,
                                             function ($constraint) {
+                                                 $constraint->aspectRatio();
                                                 $constraint->upsize();
                                             })->save(public_path($path));
 
@@ -277,26 +335,30 @@ class PostController extends Controller
      */
     public function destroy($slug)
     {  
-      $post = Post::where('slug', $slug)->with(['tags', 'imgs'])->first();
+        $post = Post::where('slug', $slug)->with(['tags', 'imgs'])->first();
 
-        if(Auth::user()->owns($post, 'author_id'))
+        if(Auth::user()->owns($post, 'author_id')) 
         {   
-            
-          
-            if($post->delete())
-            {
-                Storage::delete($post->imgs()->first()->filepath);
+            if($post->delete()) 
+            { 
+                if( count($post->imgs) )
+                {   
+                    foreach($post->imgs as $img) 
+                    Storage::delete($img->filepath);
 
-                $post->tags()->detach();
-                $post->imgs()->delete();
+                    $post->imgs()->delete();
+                }    
+                if(count($post->tags)) $post->tags()->detach();
+                
                 Session::flash('success', 'file '.$post->slug.' was successfully deleted');
                 return redirect()->route('posts.index');
+
             } else
                 return back()->withErrors('Error occured in deleting the post ');
 
 
         }  else
-             return back()->withErrors('you dont have access to delete this project');
+             return back()->withErrors('you dont have permission to delete this project');
