+++ app/Http/Controllers/User/PostController.php
@@ -7,11 +7,15 @@ use App\Http\Controllers\Controller;
 use App\Tag;
 use App\Post;
 use App\Img;
+use App\User;
+use App\Faculty;
 use Auth;
 use Image;
 use Validator;
 use Session;
 use Storage;
+use Notification;
+use App\Notifications\PostNotification;
 
 class PostController extends Controller
 {   
@@ -53,14 +57,86 @@ class PostController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
-    {   
+    {  // dd($request);
+        if($request->submit == 'Yes')
+        {  
+            Validator::make($request->all(), [
+            'content' => 'required|max:10000',
+            'tags' => 'required|array|max:20',
+            'tags.*' => 'required|max:255|string',
+            'image' => 'nullable|file|image|max:5120',
+            
+            'faculty' => 'sometimes|required|array|max:12',
+            'faculty.*' => 'required|string',
+            'start_rollno' => 'sometimes|array|max:12',
+            'start_rollno.*' =>'required_with:faculty.*|integer|min:0|max:400',
+            'end_rollno' => 'sometimes|required_with:start_rollno|array|max:12',
+            'end_rollno.*' => 'required_with:faculty.*|integer|min:0|max:400',
+            'year' => 'sometimes|required|array|max:12',
+            'year.*' => ' required_with:faculty.*|integer',
+            'ind_rollno' => 'sometimes|nullable|array|max:40',
+            'ind_rollno.*' => 'nullable|alpha_num|size:10'
+            ])->validate();
 
-        Validator::make($request->all(), [
+            $roll_no=[];
+            $college = strtoupper(session('tenant'));
+            for($i=0; $i<count($request->faculty); $i++)
+            {  
+                //to check and setup roll no's range for diffrent conditions
+                /*
+                if($request->start_rollno[$i]==NULL ||  $request->end_rollno[$i]==NULL)
+                    $roll_flag = 'all';
+                else if($request->start_rollno[$i] &&  $request->end_rollno[$i])
+                    {
+                    if($request->end_rollno[$i] < $request->start_rollno[$i])
+                       $request->end_rollno[$i]= $request->start_rollno[$i];
+                    }
+                else if($request->start_rollno[$i])
+                    $request->end_rollno[$i] = $request->start_rollno[$i];
+                else
+                     $request->start_rollno[$i]  = $request->end_rollno[$i];
+                 */
+                 if($request->end_rollno[$i] < $request->start_rollno[$i])
+                       $request->end_rollno[$i]= $request->start_rollno[$i];
+                   
+                   $faculties=[]; 
+                if($request->faculty[$i] == 'All')
+                    $faculties = Faculty::pluck('name');
+                 else
+                 $faculties[] =  Faculty::findOrFail(2)->value('name');
+                //dd($faculties);
+                for($k=0; $k<count($faculties); $k++)
+                {
+                    for($j=$request->start_rollno[$i]; $j<=$request->end_rollno[$i]; $j++ )
+                    {   
+                        $no = str_pad($j, 3, '0', STR_PAD_LEFT);
+                        $roll_no[]=$college.$no.$faculties[$k].$request->year[$i];
+                    }
+                }
+            }      
+
+            $users = User::whereIn('roll_no', $roll_no)->get();
+             
+             if($request->ind_rollno) 
+             {  
+                foreach($request->ind_rollno as $roll) 
+                    $i_rollno[] = $college.$roll;
+                
+                if($i_users = User::whereIn('roll_no', $i_rollno)->get())
+                    (count($users)) ? $users->merge($i_users) : $users = $i_users;
+                                  
+             }
+            //dd($users);
+                            
+        }else
+        {
+              Validator::make($request->all(), [
             'content' => 'required|max:10000',
             'tags' => 'required|array|max:20',
             'tags.*' => 'required|max:255|string',
             'image' => 'nullable|file|image|max:5120'
-        ])->validate();
+            ])->validate();
+        }
 
         $post= new Post;
 
@@ -71,6 +147,7 @@ class PostController extends Controller
        //inserting new tags created by user
         if($post->save())
         {  
+                
             foreach($request->tags as $tag)
             {   
                 $string=str_split($tag);
@@ -108,8 +185,9 @@ class PostController extends Controller
                     Storage::makeDirectory($dir, 0775, true);
                 }
 
-               Image::make($image)->resize(600, 350,
+               Image::make($image)->resize(720, 720,
                                         function ($constraint) {
+                                           $constraint->aspectRatio();
                                             $constraint->upsize();
                                         })->save(public_path($path));
 
@@ -120,17 +198,25 @@ class PostController extends Controller
                 $img=new Img;
                 $img->filepath=$path;
                
-                //$img->save();          
-               //dd($img);
                  if( !($post->imgs()->save($img)) )
                     $msg = '. Also Error in uploading image.';
             }
 
 
             if($post->tags()->sync($tag_ids, false))
-            {
+            {   
+            
                 Session::flash('success', 'successfully added new post to discussions and forums'.$msg);
-                return redirect()->route('user.posts.show', $post->slug);
+
+                if($request->submit == 'Yes')
+                {
+                  $when = now()->addSeconds(10);
+                  $data['owner'] = Auth::user()->name;
+                  $data['url'] = route('user.posts.show', $post->slug);
+                  Notification::send($users->unique(), (new PostNotification($post, $data))->delay($when));
+                }
+            
+               return redirect()->route('user.posts.show', $post->slug);
             }
             else
                 return back()->withErrors('problem  adding tags to the post '.$msg);
@@ -149,11 +235,8 @@ class PostController extends Controller
     public function show($slug)
     {
         $post = Post::where('slug', $slug)->with('tags')->first();
-        //dd($post);
-        /*
-        $post1 = Post::where('slug', $slug)->get();
-        dd($post1->imgs);*/
-            $post->increment('view_count');
+        abort_if(!$post, 404);
+        $post->increment('view_count');
         return view('user.posts.show', ['post' => $post]);
     }
 
@@ -165,7 +248,7 @@ class PostController extends Controller
      */
     public function edit($slug)
     {
-        $post = Post::where('slug', $slug)->with('tags')->first();
+        $post = Post::where('slug', $slug)->with(['tags', 'imgs'])->first();
         if(Auth::user()->owns($post, 'author_id'))
         { 
             $tags = Tag::all();
@@ -276,27 +359,30 @@ class PostController extends Controller
      */
     public function destroy($slug)
     {  
-      $post = Post::where('slug', $slug)->with(['tags', 'imgs'])->first();
+        $post = Post::where('slug', $slug)->with(['tags', 'imgs'])->first();
 
-        if(Auth::user()->owns($post, 'author_id'))
+        if(Auth::user()->owns($post, 'author_id')) 
         {   
-            
-          
-            if($post->delete())
-            {
-                Storage::delete($post->imgs()->first()->filepath);
+            if($post->delete()) 
+            { 
+                if( count($post->imgs) )
+                {   
+                    foreach($post->imgs as $img) 
+                    Storage::delete($img->filepath);
 
-                $post->tags()->detach();
-                $post->imgs()->delete();
+                    $post->imgs()->delete();
+                }    
+                if(count($post->tags)) $post->tags()->detach();
+                
                 Session::flash('success', 'file '.$post->slug.' was successfully deleted');
                 return redirect()->route('user.posts.index');
+
             } else
                 return back()->withErrors('Error occured in deleting the post ');
 
 
         }  else
-             return back()->withErrors('you dont have access to delete this project');
-
+             return back()->withErrors('you dont have permission to delete this project');       
 
         
     }
